refactor(core): move debug domain into editor
