--- src/core/contextual-ui/usecases/index.tsx
@@ -1,1 +1,0 @@
-export * from './SetDebuggerEntityUseCase';
