+++ src/core/editor/domain/entities/EntityOtuline.tsx
