+++ src/core/editor/domain/entities/EntityOutline.test.tsx
