+++ src/core/editor/domain/entities/IDebuggerEntity.tsx
