+++ src/core/editor/domain/entities/Pivot.test.tsx
