+++ src/core/editor/domain/entities/index.tsx
