+++ src/core/editor/infrastructure/adapters/ColorObjectPicker.test.tsx
@@ -1,8 +1,8 @@
 import { BlendMethod, CanvasDevice, DOMImageLoader, DrawPrimitiveCommand, GameObject, PrimitiveType, RenderCommand, Renderer, SetBlendingMethodCommand, SetTransformMatrixCommand, ShapeComponent } from "sparkengineweb";
 import { ColorObjectPicker } from "./ColorObjectPicker";
 import { uuidToRgb } from "../../../common/uuidToRgb";
-import Pivot from "../../../contextual-ui/Pivot";
-import { EntityOutline } from "../../../contextual-ui";
+import Pivot from "../../domain/entities/Pivot";
+import { EntityOutline } from "../../domain/entities";
 
 class RendererTestDouble extends Renderer {
     public renderCommands: RenderCommand[] = [];
