+++ src/core/editor/infrastructure/adapters/ColorObjectPicker.tsx
@@ -2,8 +2,8 @@ import { BlendMethod, CanvasDevice, DrawPrimitiveCommand, IEntity, ImageLoader, 
 import { uuidToRgb } from "../../../common/uuidToRgb";
 import { Factory, Optional } from "../../../common/";
 import { ObjectPicker } from "../../domain/ports/ObjectPicker";
-import Pivot from "../../../contextual-ui/Pivot";
-import { EntityOutline } from "../../../contextual-ui";
+import Pivot from "../../domain/entities/Pivot";
+import { EntityOutline } from "../../domain/entities";
 
 export class ColorObjectPicker extends RenderSystem implements ObjectPicker {
     private colorToEntityMap: Map<string, IEntity> = new Map();
